+++ docs/docs/.vuepress/components/GLTFLoaderFunction.vue
@@ -9,7 +9,9 @@ const created = async (scene, { camera }) => {
   camera.position.set(0, 1.5, 3)
 
   // Load model to scene.
-  const model = await GLTFLoader('/models/girl.glb')
+  const model = await GLTFLoader('/models/girl.glb', true, (progress: number) => {
+    console.log('progress', progress)
+  })
   scene.add(model)
 }
 </script>