+++ src/types/global.d.ts
@@ -1,5 +1,9 @@
 declare global {
   type Position = [number, number, number]
+  type LoadEvent = {
+    type: string
+    progress: number
+  }
 }
 
 export {}